chore: Model 수정
