+++ FourStars/FourStars/Source/Model/AfternoonHabit.swift
@@ -1,5 +1,5 @@
 //
-//  ExtraHabit.swift
+//  AfternoonHabit.swift
 //  FourStars
 //
 //  Created by Soop on 4/15/25.
@@ -9,17 +9,18 @@ import Foundation
 import SwiftData
 
 @Model
-class ExtraHabit: HabitDetail {
+class AfternoonHabit: HabitDetail {
     var uuid: String = UUID().uuidString
-    var timeSlot: TimeSlot = TimeSlot.morning
-    var habit: ExtraHabitType?
     
+    var timeSlot: TimeSlot = TimeSlot.evening
+    var habit: AfternoonHabitType?
+
     var quantity: Int?
     var startTime: Date?
     var endTime: Date?
-    
+       
     init(
-        habit: ExtraHabitType? = nil,
+        habit: AfternoonHabitType? = nil,
         quantity: Int? = nil,
         startTime: Date? = nil,
         endTime: Date? = nil
@@ -32,14 +33,14 @@ class ExtraHabit: HabitDetail {
     }
 }
 
-enum ExtraHabitType: String, CaseIterable, Identifiable, Codable {
-    case gratitude = "감사 표현"
-    case noPhone = "폰 사용 줄이기"
-    case deepBreath = "심호흡"
-
+enum AfternoonHabitType: String, CaseIterable, Identifiable, Codable {
+    case walk = "산책"
+    case read = "책"
+    case meditate = "명상"
+    
     var id: String { rawValue }
 }
 
-extension ExtraHabit {
-    static var stub01: ExtraHabit = .init(habit: .deepBreath, quantity: nil, startTime: nil, endTime: nil)
+extension AfternoonHabit {
+    static var stub01: AfternoonHabit = .init(habit: .meditate, quantity: nil, startTime: nil, endTime: nil)
 }
