+++ FourStars/FourStars/Source/Model/ExtraHabit.swift
@@ -1,5 +1,5 @@
 //
-//  EveningHabit.swift
+//  ExtraHabit.swift
 //  FourStars
 //
 //  Created by Soop on 4/15/25.
@@ -9,17 +9,18 @@ import Foundation
 import SwiftData
 
 @Model
-class EveningHabit: HabitDetail {
+class ExtraHabit: HabitDetail {
     var uuid: String = UUID().uuidString
     var timeSlot: TimeSlot = TimeSlot.morning
-    var habit: EveningHabitType?
+    
+    var habit: ExtraHabitType?
     
     var quantity: Int?
     var startTime: Date?
     var endTime: Date?
     
     init(
-        habit: EveningHabitType? = nil,
+        habit: ExtraHabitType? = nil,
         quantity: Int? = nil,
         startTime: Date? = nil,
         endTime: Date? = nil
@@ -32,14 +33,14 @@ class EveningHabit: HabitDetail {
     }
 }
 
-enum EveningHabitType: String, CaseIterable, Identifiable, Codable {
-    case reflect = "하루 회고"
-    case journal = "일기 쓰기"
-    case tidyUp = "정리 정돈"
+enum ExtraHabitType: String, CaseIterable, Identifiable, Codable {
+    case gratitude = "감사 표현"
+    case noPhone = "폰 사용 줄이기"
+    case deepBreath = "심호흡"
 
     var id: String { rawValue }
 }
 
-extension EveningHabit {
-    static var stub01: EveningHabit = .init(habit: .journal, quantity: nil, startTime: nil, endTime: nil)
+extension ExtraHabit {
+    static var stub01: ExtraHabit = .init(habit: .deepBreath, quantity: nil, startTime: nil, endTime: nil)
 }
