+++ FourStars/FourStars/Source/Type/TimeSlot.swift
@@ -21,4 +21,17 @@ enum TimeSlot: String, Codable, Identifiable, CaseIterable {
             case .extra: return "기타"
         }
     }
+    
+//    var startTime: Date {
+//        switch self {
+//        case .morning:
+//            <#code#>
+//        case .afternoon:
+//            <#code#>
+//        case .evening:
+//            <#code#>
+//        case .extra:
+//            <#code#>
+//        }
+//    }
 }
